+++ src/bk-user/bkuser/apps/data_source/constants.py
@@ -8,25 +8,27 @@ Unless required by applicable law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-from enum import auto
+from enum import auto, Enum
 
 from django.utils.translation import gettext_lazy as _
 
-from bkuser.common.enum import AutoLowerEnum
 
+class DataSourceType(Enum):
 
-class DataSourceType(AutoLowerEnum):
-    """目录类型
-
-    TODO: 当目录和数据源解耦完成吼，这里的类型实际上就应该去除
-    """
-
-    LOCAL = auto()
-    MAD = auto()
-    LDAP = auto()
+    LOCAL = "local"
+    MAD = "mad"
+    LDAP = "ldap"
 
     _choices_labels = (
         (LOCAL, _("本地目录")),
         (MAD, _("Microsoft Active Directory")),
         (LDAP, _("OpenLDAP")),
     )
+
+    @classmethod
+    def get_choices(cls) -> tuple:
+        return cls._choices_labels.value
+
+
+DEFAULT_DATA_SOURCE_NAME = "本地数据源"
+
